requirements.txt
face-recognition-models==0.3.0
numpy==1.18.5
opencv-contrib-python==4.2.0.34
